Create a reusable MilestonesScaffold out of common widgets
