+++ lib/screens/authenticate/authenticate.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:milestones/screens/authenticate/register.dart';
+import 'package:milestones/screens/authenticate/register/register.dart';
 
-import 'login.dart';
+import 'login/login.dart';
 
 class Authenticate extends StatefulWidget {
   const Authenticate({Key? key}) : super(key: key);
@@ -19,13 +19,8 @@ class _AuthenticateState extends State<Authenticate> {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.lightGreenAccent,
-      body: SafeArea(
-        child: showSignIn
-            ? LoginScreen(toggleView: _toggleView)
-            : RegisterScreen(toggleView: _toggleView),
-      ),
-    );
+    return showSignIn
+        ? LoginScreen(toggleView: _toggleView)
+        : RegisterScreen(toggleView: _toggleView);
   }
 }
